create resonance module and algorithms here from molecule module
same for unit tests
